finalizado views, ta com bug

+++ laravel/database/migrations/2019_10_27_190941_create_produtos_table.php
@@ -17,7 +17,7 @@ class CreateProdutosTable extends Migration
             $table->bigIncrements('id');
             $table->string('name');
             $table->string('description');
-            $table->decimal('value');
+            $table->integer('value');
             $table->timestamps();
         });
     }
